Rx: provide global error handler.
This will provide a global error handler for Rx exceptions, which might
come from either:

A) Observables after they have been disposed, in which case we don't
want the exception anyway.

B) Observables that don't have an onError() handler in their Subscriber
intentionally, in which case we don't want the exception anyway.

C) Observables that don't have an onError() handler in their Subscriber
unintentionally, in which case incorrect behavior will likely manifest
itself, which will lead us to provide an onError() handler to the
Subscriber.

Considering all of these cases, it seems logical and safe to provide an
empty consumer as the global Rx error handler. This means, however, that
we'll need to be extra mindful about providing onError() handlers to all
our Subscribers, or omitting them with deliberate knowledge that the
errors will be suppressed.

Change-Id: I280e00cd297a20f5b96157b39d8b8e8d448457a6