Priority fix: improve error/status handling in WebView XHR callback.
Recent versions of Chromium (at least on Android) have a peculiar behavior
where it will send a state-change event with a `status` of zero (0), which
is not a valid HTTP status code, usually to indicate that the request
timed out, or failed for other reasons. We were handling this case
incorrectly, which was leading to an error being propagated back to the
Java layer.

This also expands the range of acceptable HTTP status codes that indicate
"success", since it's not just 200 that can be treated as a successful
response.

Change-Id: I2171677825e234c9c1f52b4dfe3d330d3595a119