Fix possible memory leak upon exit transition.
This introduces a more robust way of tracking whether we have MainActivity
in our backstack. If we do, then the transition *back* from PageActivity
should be overridden. And if we don't, then we can use the new constructed
shared-element transition.

This also improves the actual logic of navigating back and forth between
MainActivity and PageActivity, taking all the associated transitions into
account.

And this also implicitly fixes a platform-level memory leak, whereby we
were overriding the pending transition while *also* providing a shared-
element bundle.

Change-Id: Ic4fe1b2a3aa9fd401623b726c6264804131f35f4