Make isOnline() function a lot less expensive.
This patch moves the isOnline() method to the global
NetworkConnectivityReceiver, and basically caches the online state. Of
course, when the receiver gets an explicit network change event, the
online state will be updated too.

This makes the function of checking whether the device is online a whole
lot more efficient.

Change-Id: Id268ade13e3ad14bbd410506e66fca0cc00eb501