Don't cache images when opening page in background tab.
The PageCacher class is used for pre-fetching the page if it's opened in a
background tab, so that when the user goes to that tab, the page will load
faster. However, all the images in that page don't actually need to be
cached, especially because we now have lazy-loading of images.

Change-Id: I66dd41bbe3ffc6c11ff215d1e60df7678e4d382e