Fix applying correct theme to articles.
After the recent update to reload the stub HTML on every page load, the
theme of the article was no longer being applied correctly on subsequent
loads.

This corrects the setting of the current theme over the Bridge upon every
page reset.

Also, this fixes the issue of the WebView temporarily being fully white
(even when the theme is dark) by explicitly setting the background color
of the stub HTML to correspond to the current theme.

Change-Id: I088896287b60414d0512100bedea4e1d6ae5db08