Don't run the notification polling service unnecessarily.
The notification polling service is set up to run unconditionally, and then
determines whether to poll the API based on whether the user is logged in,
and polling is enabled.

Even though the service itself doesn't consume any resources when it runs
without polling, it still creates a nuisance wakeup which gets logged by the
system and appears in our statistics as an undesirable ratio of wakeups.

Change-Id: I52f098427f42c0a6f282e89a60ba4f8bca0eea48