Edit action feed provider.
This provides an interface to receive editing suggestions for articles
that don't have a description (in the current wiki language), as well as
articles that have a description in the current language, but are missing
a description in a target language.

Please use the handy-dandy new Dev options to test it.

Change-Id: I4bb3d8a855934d172b06f59dd197dd9efe74fc45