Fix/improve persistence of scroll position when browsing articles.
You might have noticed that the app doesn't always preserve the exact
scroll position when you leave and come back to an article. (or at least
it's not totally reliable.)

The logic for preserving/restoring the scroll position was a little
rickety, and needed an update. This patch simplifies the logic, and brings
it fully into native code, instead of relying on the WebView to execute
the scrolling.

Change-Id: I7f2703a963e282d8275b46f43d7582948fdd5ac8