Add dialog that's shown when editor tasks list is unlocked.
Some notes:
This merely adds the logic to show the unlock dialog when the user makes
their 50th edit.  It also adds the logic to keep track of reverted edits,
but does not yet tie the two bits together, while we finalize the actual
behavior that we want to see here.

Bug: T207333
Change-Id: I77d092cf9b20b9e09a43d44d08db243830066bbf