Fix/optimize switching languages when translating descriptions.
There was a possible race condition when selecting a different "from"
language that would cause the MissingDescriptionProvider to look for
articles with the same "from" and "to" languages, which would cause it
to be an infinite loop.

This also optimizes the MissingDescription logic to request a single
Summary instead of two Summaries, since we were using the "source"
Summary only to get the description, but we actually already have the
source description from the Wikidata call.

Change-Id: If05aee4c08aab3c7b7eaf172dc7b8fe7a7d89406