Modernize handling of errors from the API.
The format of the error structures that we were expecting from the API
was actually the "legacy" format, and will eventually be deprecated.

We need to adapt our model classes to accept the most current error
structure, which contains improvements for internationalization and
better flexibility.  This patch updates the expected structure, and
updates our test fixtures to match the new format.

Change-Id: I225fea472d1de109fbe0766707a4e041d004396f