MEP (#1774)
* Add Event Platform Client

Adds the client code for interacting with the Wikimedia Event Platform
(a.k.a. "Modern Event Platform").

The bulk of the new implementation is added in the new package
org.wikipedia.analytics.eventplatform. Conceptually, the code is
structured as a plain Java client "library" (represented by the
EventPlatformClient class) and an EventPlatformClientIntegration class
through which the EventPlatformClient "library" class interacts with
various APIs provided by the Android platform and the app itself.

The remaining classes in the org.wikipedia.analytics.eventplatform
package are mostly model classes, for events (Event), event platform
intake service responses (EventServiceResponse), stream configurations
(StreamConfig), and the sampling configuration components of stream
configurations (SamplingConfig). EventService is a Retrofit service
interface for the Event Platform intake APIs. The
DestinationEventService enum contains the configurations for possible
destination APIs designated in event stream configurations. (N.B.
To match the iOS client behavior, the request for stream
configurations filters for those with the destination event service
eventgate-analytics-external, and accordingly only that service will
be used. The app can be updated later to remove the filtering on
stream config requests in the event that it is desired to submit app
client error events to the Event Platform.

Instruments added to the code should interact with the
EventPlatformClient class, specifically through the static method
EventPlatformClient.submit.

In WikipediaApp.onCreate(), the Event Platform Client is initialized.
The client requests stream configurations from SharedPreferences, if
available, and makes a request to the MediaWiki API for updated
configurations. When an instrument submits an event to the client via
EventPlatformClient.submit(), the client searches for an applicable
stream configuration. If one is found, the client applies any
sampling logic directed by the stream configuration. If the event is
in sample, the client supplements the event with a timestamp, if
needed, and enqueues the event for batched submission to the
appropriate event platform intake API. Events are enqueued and
submitted periodically in batches to limit the frequency with which
the app wakes the device radio for event logging. If the queue reaches
a configured maximum size, it is flushed regardless of the length of
time that has passed since the previous event submission.

A developer preference has been added to send a test event to an
Event Platform intake API. The next steps are to identify a schema
used by the Android app to migrate to the Event Platform, do the
migration, and update the app to submit events to the migrated
schema.

Since this commit is already large, session timeouts are left for a
follow-up PR.

Bug: T228179

* EventPlatformClient class structure changes (#1844)

* EventPlatformClient does not need to be a singleton class

* checkstyle

* Purge the timer

* synchronize the crucial methods

* Drop usage of BigInteger and BigDecimal

The BigInteger and BigDecimal classes carry a performance penalty and
are overkill for constructing and manipulating 80-bit sampling
identifiers.

* Use buffering only whilst Offline (#1805)

* Have smaller batch sizes when online, and a larger buffer whilst offline

* Correction to static variable

* Batching the events

* set-right the tests

* Add a null check

* CR comments addressed

* arranges events into stream map

* Correction to naming

* Enabled checck is redundant

* Create analytics service through ServiceFactory. (#1818)

* Create an analytics service through the ServiceFactory.

* Change in wiki usage in analytics service

* CR commenmt

* Minor code change

* Optimize MEP random id generation.

Change-Id: I5cabdc9716ff6dfb68ca4ae8585d44144e320a81

* Queuing issues resolved

* Unused code

* Correction to enabled

* fixing build

* Updated command format

* minor update

* Clear disposables

* Checkstyle fix

* Fix build

* Remove test

* Removed unnecessary app instance

* Removed unnecessary check

Co-authored-by: Michael Holloway <mholloway@wikimedia.org>
Co-authored-by: Dmitry Brant <dbrant@wikimedia.org>
Co-authored-by: Cooltey Feng <cfeng@wikimedia.org>