MEP: Send events immediately for testing (#2369)
* Remove stream name regex-matching

EventStreamConfigs supports defining rules for streams with names that
match (JavaScript) regular expressions. However, this is only used in
practice for job-queue-related jobs, and the app need not support it.

* Send events immediately for testing

Sends events immediately upon submission to the MEP client to match
the legacy eventlogging behavior. We'll compare the resulting events
with those sent via the legacy system in order to determine
conclusively whether events are being leaked in the queueing process.

Phab: https://phabricator.wikimedia.org/T281001

Co-authored-by: Sharvani Haran <sharvaniharan@users.noreply.github.com>