Support exporting history to csv
On the "Search" section, the application shows the history of the pages
the user has viewed. The history view supports filtering and deleting.

In this commit, we add a new option to "export" the history via a csv
file. The CSV file contains all the data from the history table. After
the file is created, the user is prompted on how they want to share the
file.

Bug: T249340