Integrate with Phabricator message bot.
This integrates with our Phab bot which should automatically post messages on the corresponding task.
From now on, we can use the old-style GerritBot convention of mentioning the task at the bottom of the pull request description (by saying Bug: T...), and it will be posted automatically to the corresponding task.

Bug: T360680