Support URL sharing on recent screen for images
Similar to https://github.com/wikimedia/apps-android-wikipedia/pull/4816 but for pictures.

Unlike the name, .commonsUrl, it works for both images uploaded on
Wikimedia Commons and locally uploaded images.

To test this open an article, one an image from, hit the rectangle
button above the emulator, click on the anchor icon on the recents
screen.